+++ admin/admin.php
@@ -53,9 +53,15 @@ class WPSubtitle_Admin {
 
 		// Setup Field / Meta Box
 		if ( WPSubtitle::is_supported_post_type( $post_type ) ) {
-			if ( self::edit_form_after_title_supported( $post_type ) ) {
+
+			$position = self::subtitle_field_position( $post_type );
+
+			if ( 'after_title' == $position ) {
 				add_action( 'admin_head', array( 'WPSubtitle_Admin', '_add_admin_styles' ) );
 				add_action( 'edit_form_after_title', array( 'WPSubtitle_Admin', '_add_subtitle_field' ) );
+			} elseif ( 'before_title' == $position ) {
+				add_action( 'admin_head', array( 'WPSubtitle_Admin', '_add_admin_styles' ) );
+				add_action( 'edit_form_top', array( 'WPSubtitle_Admin', '_add_subtitle_field' ) );
 			} else {
 				add_action( 'add_meta_boxes', array( 'WPSubtitle_Admin', '_add_meta_boxes' ) );
 			}
@@ -63,6 +69,7 @@ class WPSubtitle_Admin {
 			add_filter( 'manage_edit-' . $post_type . '_columns', array( 'WPSubtitle_Admin', 'manage_subtitle_columns' ) );
 			add_action( 'manage_' . $post_type . '_posts_custom_column', array( 'WPSubtitle_Admin', 'manage_subtitle_columns_content' ), 10, 2 );
 			add_action( 'quick_edit_custom_box', array( 'WPSubtitle_Admin', 'quick_edit_custom_box' ), 10, 2 );
+
 		}
 
 	}
@@ -110,9 +117,24 @@ class WPSubtitle_Admin {
 
 		$new_columns = array();
 
+		// Subtitle column after...
+		$after_column = '';
+		if ( array_key_exists( 'title', $columns ) ) {
+			$after_column = 'title';
+		} elseif ( array_key_exists( 'name', $columns ) ) {
+			$after_column = 'name';
+		}
+
+		// Add column
+		if ( empty( $after_column ) ) {
+			$columns['wps_subtitle'] = __( 'Subtitle', 'wp-subtitle' );
+			return $columns;
+		}
+
+		// Insert column
 		foreach ( $columns as $column => $value ) {
 			$new_columns[ $column ] = $value;
-			if ( 'title' == $column ) {
+			if ( $after_column == $column ) {
 				$new_columns['wps_subtitle'] = __( 'Subtitle', 'wp-subtitle' );
 			}
 		}
@@ -253,7 +275,11 @@ class WPSubtitle_Admin {
 	public static function _add_meta_boxes() {
 		$post_types = WPSubtitle::get_supported_post_types();
 		foreach ( $post_types as $post_type ) {
-			add_meta_box( 'wps_subtitle_panel',  self::get_meta_box_title( $post_type ), array( 'WPSubtitle_Admin', '_add_subtitle_meta_box' ), $post_type, 'normal', 'high' );
+
+			$positiom = self::gutenberg_supported( $post_type ) ? 'side' : 'normal';
+
+			add_meta_box( 'wps_subtitle_panel',  self::get_meta_box_title( $post_type ), array( 'WPSubtitle_Admin', '_add_subtitle_meta_box' ), $post_type, $positiom, 'high' );
+
 		}
 	}
 
@@ -437,6 +463,48 @@ class WPSubtitle_Admin {
 			return false;
 		}
 		return ! apply_filters( 'wps_subtitle_use_meta_box', false, $post_type );
+	}
+
+	/**
+	 * Gutenberg Supported
+	 *
+	 * @since  3.1
+	 *
+	 * @param   string  $post_type  Post type.
+	 * @return  bool
+	 */
+	private static function gutenberg_supported( $post_type = '' ) {
+
+		if ( function_exists( 'gutenberg_can_edit_post_type' ) && gutenberg_can_edit_post_type( $post_type ) && ! isset( $_GET['classic-editor'] ) ) {
+			return true;
+		}
+
+		return false;
+
+	}
+
+	/**
+	 * Subtitle Field Position
+	 *
+	 * @since  3.1
+	 *
+	 * @param  string  $post_type  Post type.
+	 * @param  string              Position.
+	 */
+	private static function subtitle_field_position( $post_type = '' ) {
+
+		$position = apply_filters( 'wps_subtitle_field_position', 'after_title', $post_type );
+
+		if ( self::gutenberg_supported( $post_type ) ) {
+			return '';
+		}
+
+		if ( ! self::edit_form_after_title_supported( $post_type ) ) {
+			return '';
+		}
+
+		return $position;
+
 	}
 
 }